refactor(connectors): reorganize connectors
