+++ packages/connector-aliyun-sms/src/types.ts
@@ -1,5 +1,7 @@
 import { z } from 'zod';
 
+import { SmsTemplateType } from './constant';
+
 export type { Response } from 'got';
 
 export type SendSmsResponse = { BizId: string; Code: string; Message: string; RequestId: string };
@@ -26,21 +28,6 @@ export interface PublicParameters {
   SignatureVersion?: string;
   Timestamp?: string;
   Version?: string;
-}
-
-/**
- * Details of SmsTemplateType can be found at:
- * https://next.api.aliyun.com/document/Dysmsapi/2017-05-25/QuerySmsTemplateList.
- *
- * For our use case is to send passcode sms for passwordless sign-in/up as well as
- * reset password, the default value of type code is set to be 2.
- */
-enum SmsTemplateType {
-  Notification = 0,
-  Promotion = 1,
-  Passcode = 2,
-  InternationalMessage = 6,
-  PureNumber = 7,
 }
 
 /**
